requirements.txt
Babel==2.4.0
BzrTools==2.6.0
Flask==0.10.1
Flask-Babel==0.11.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
Flask-WhooshAlchemy==0.56
Jinja2==2.7.2
MarkupSafe==0.18
Pillow==2.3.0
PyJWT==1.5.0
PySocks==1.6.7
SQLAlchemy==1.1.10
SecretStorage==2.0.0
Tempita==0.5.2
WTForms==2.1
Werkzeug==0.9.4
Whoosh==2.7.4
argparse==1.2.1
bcrypt==3.1.3
blinker==1.3
bzr==2.7.0dev1
cffi==1.10.0
chardet==2.0.1
colorama==0.2.5
configobj==4.7.2
coverage==4.4.1
decorator==3.4.0
duplicity==0.6.23
flipflop==1.0
guess-language==0.2
gunicorn==19.7.1
heroku==0.1.4
html5lib==0.999
httplib2==0.8
ipython==1.2.1
itsdangerous==0.22
keyring==3.5
launchpadlib==1.10.2
lazr.restfulclient==0.13.3
lazr.uri==1.0.3
lockfile==0.8
lpthw.web==1.1
matplotlib==1.3.1
mercurial==2.8.2
numpy==1.8.2
oauth==1.0.1
paramiko==1.10.1
pbr==3.0.1
pexpect==3.1
pyOpenSSL==0.13
pycparser==2.17
pycrypto==2.6.1
pygobject==3.12.0
pygpgme==0.3
pyinotify==0.9.4
pyparsing==2.0.1
python-apt==0.9.3.5ubuntu2
python-dateutil==1.5
python-openid==2.2.5
pytz==2012c
requests==2.2.1
scipy==0.13.3
simplegeneric==0.8.1
simplejson==3.3.1
six==1.10.0
sqlalchemy-migrate==0.11.0
sqlparse==0.2.3
stevedore==0.14.1
twilio==6.2.0
urllib3==1.7.1
virtualenv==1.11.4
virtualenv-clone==0.2.4
virtualenvwrapper==4.1.1
wadllib==1.3.2
wheel==0.24.0
wsgiref==0.1.2
zope.interface==4.0.5
